2020.9.11

+++ home-care-center-start/src/main/java/edu/hust/controller/AreaController.java
@@ -4,12 +4,16 @@ import edu.hust.common.constant.ApiCodeEnum;
 import edu.hust.common.util.RandomUUID;
 import edu.hust.common.vo.ApiResult;
 import edu.hust.dao.dto.Area;
+import edu.hust.dao.dto.Room;
+import edu.hust.dao.dto.Worker;
 import edu.hust.monitor.Monitor;
+import edu.hust.service.domain.WorkerFull;
 import edu.hust.service.service.AreaService;
 import edu.hust.common.exception.GlobalException;
+import edu.hust.service.service.RoomService;
+import edu.hust.service.service.WorkerService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import org.apache.commons.lang3.ObjectUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,6 +32,12 @@ public class AreaController {
 
     @Autowired
     private AreaService areaService;
+
+    @Autowired
+    private RoomService roomService;
+
+    @Autowired
+    private WorkerService workerService;
 
     @Autowired
     private RandomUUID randomUUID;
@@ -95,12 +105,32 @@ public class AreaController {
             @RequestParam(value = "title", required = false) String title
     ) {
         if (id == null && title == null) {
+            List<Room> roomList = roomService.getRoomList();
+            if (!roomList.isEmpty()) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
+            List<WorkerFull> workerFullList = workerService.getWorkerList();
+            for (WorkerFull workerFull : workerFullList) {
+                if (workerFull.getArea() != null) {
+                    throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+                }
+            }
             areaService.deleteAllArea();
             return ApiResult.buildSuccess();
         }else if (id == null) {
+            Area area = areaService.getAreaInfoByTitle(title);
+            if (area == null) {
+                return ApiResult.buildSuccess();
+            }
+            if (!legalDelete(area.getId())) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             areaService.deleteAreaByTitle(title);
             return ApiResult.buildSuccess();
         }else if (title == null) {
+            if (!legalDelete(id)) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             areaService.deleteAreaById(id);
             return ApiResult.buildSuccess();
         }else {
@@ -111,6 +141,22 @@ public class AreaController {
     private boolean legal(Area area) {
         if (area.getAreaTitle() == null) {
             return false;
+        }
+        return true;
+    }
+
+    private boolean legalDelete(String areaId) {
+        List<Room> roomList = roomService.getRoomList();
+        for (Room room : roomList) {
+            if (room.getArea() != null && areaId == room.getArea().getId()) {
+                return false;
+            }
+        }
+        List<WorkerFull> workerFullList = workerService.getWorkerList();
+        for (WorkerFull workerFull : workerFullList) {
+            if (workerFull.getAreaId() == areaId) {
+                return false;
+            }
         }
         return true;
     }
