+++ home-care-center-start/src/main/java/edu/hust/controller/DishController.java
@@ -4,15 +4,20 @@ import edu.hust.common.constant.ApiCodeEnum;
 import edu.hust.common.util.RandomUUID;
 import edu.hust.common.vo.ApiResult;
 import edu.hust.dao.dto.Dish;
+import edu.hust.dao.dto.DishSet;
 import edu.hust.monitor.Monitor;
+import edu.hust.service.domain.DishSetFull;
 import edu.hust.service.service.DishService;
 import edu.hust.common.exception.GlobalException;
+import edu.hust.service.service.DishSetService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -28,6 +33,9 @@ public class DishController {
 
     @Autowired
     private DishService dishService;
+
+    @Autowired
+    private DishSetService dishSetService;
 
     @Autowired
     private RandomUUID randomUUID;
@@ -95,12 +103,26 @@ public class DishController {
             @RequestParam(value = "name", required = false) String name
     ) {
         if (id == null && name == null) {
+            List<DishSetFull> dishSetFullList = dishSetService.getDishSetList();
+            if (!dishSetFullList.isEmpty()) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             dishService.deleteAllDish();
             return ApiResult.buildSuccess();
         } else if (id == null) {
+            Dish dish = dishService.getDishByName(name);
+            if (dish == null) {
+                return ApiResult.buildSuccess();
+            }
+            if (!legalDelete(dish.getId())) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             dishService.deleteDishByName(name);
             return ApiResult.buildSuccess();
         } else if (name == null) {
+            if (!legalDelete(id)) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             dishService.deleteDishById(id);
             return ApiResult.buildSuccess();
         } else {
@@ -111,6 +133,27 @@ public class DishController {
     private boolean legal(Dish dish) {
         if (dish.getName() == null) {
             return false;
+        }
+        return true;
+    }
+
+    private boolean legalDelete(String dishId) {
+        List<DishSetFull> dishSetFullList = dishSetService.getDishSetList();
+        for (DishSetFull dishSetFull : dishSetFullList) {
+            List<String> dishList = new ArrayList<>();
+            dishList.add(dishSetFull.getMon());
+            dishList.add(dishSetFull.getTue());
+            dishList.add(dishSetFull.getWed());
+            dishList.add(dishSetFull.getThu());
+            dishList.add(dishSetFull.getFri());
+            dishList.add(dishSetFull.getSat());
+            dishList.add(dishSetFull.getSun());
+            String ids = String.join(",", dishList);
+            String str[] = ids.split(",");
+            List<String> idList = Arrays.asList(str);
+            if (idList.contains(dishId)) {
+                return false;
+            }
         }
         return true;
     }
